requirements.txt
annotated-types==0.7.0
bert-extractive-summarizer==0.10.1
blinker==1.8.2
blis==0.7.11
cached-property==1.5.2
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
coverage==5.2.1
coveralls==1.11.1
cymem==2.0.8
docopt==0.6.2
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
frozendict==2.4.4
fsspec==2024.6.0
genanki==0.13.1
httpretty==1.1.4
huggingface-hub==0.23.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mock==3.0.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pipeline==0.1.0
pipelines==0.0.1
preshed==3.0.9
protobuf==5.27.1
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12.1
thinc==8.2.4
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.1
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
youtube-transcript-api==0.6.2
zipp==3.19.2
